+++ test/mocks/hashable_mock.dart
@@ -1,15 +1,17 @@
-// ignore_for_file: avoid_equals_and_hash_code_on_mutable_classes
-
+import 'package:flutter/foundation.dart';
 import 'package:folly_fields/util/hashable.dart';
 
 ///
 ///
 ///
-class HashableMock with Hashable {
-  String name = 'name';
-  int age = 20;
-  List<String> stringList = <String>['A'];
-  Map<String, String> stringMap = <String, String>{'A': 'a'};
+@immutable
+class HashableObjectMock with Hashable {
+  final String name = 'name';
+  final int age = 20;
+  final List<String> stringList = <String>['A'];
+  final Map<String, String> stringMap = <String, String>{'A': 'a'};
+  final bool boolean = true;
+  final Object? nullObject = null;
 
   ///
   ///
@@ -20,6 +22,8 @@ class HashableMock with Hashable {
     map['age'] = age;
     map['stringList'] = stringList;
     map['stringMap'] = stringMap;
+    map['boolean'] = boolean;
+    map['nullObject'] = nullObject;
 
     return map;
   }
@@ -35,4 +39,35 @@ class HashableMock with Hashable {
   ///
   @override
   bool operator ==(Object other) => hashCode == other.hashCode;
+}
+
+///
+///
+///
+@immutable
+class HashableIterableMock<T> with Hashable {
+  final Iterable<T?> iterable;
+
+  ///
+  ///
+  ///
+  HashableIterableMock(this.iterable);
+
+  ///
+  ///
+  ///
+  @override
+  int get hashCode => hashIterable(iterable);
+
+  ///
+  ///
+  ///
+  @override
+  bool operator ==(Object other) => hashCode == other.hashCode;
+
+  ///
+  ///
+  ///
+  @override
+  String toString() => iterable.toString();
 }